(ciaopp) helper package to write domains (see below)
This is a (possibly temporary) package to write AI domains for CiaoPP.
Right now it only checks that the domain hooks are normalized (regular
names, same variable order), but it will generate those hooks (at
least until we use traits and we can check that multifile predicates
do not add any significant overhead). Note that the checks for hooks
are disabled by default (since some domain operations are still not
normalized).

Src-commit: 373f4126296bc57dff55063dce9ee14162a67176