(ciaopp) multifile hooks for domain operations (see details below)
This commit moves domain implementation declarations from
domains_hook.pl to their respective modules. In principle the domains
are truly modular now.

The interfaces are implemented with the 'aidomain' package, which uses
a custom syntax (that is undocumented and which still needs to be
improved). To minimize the changes in the rest of CiaoPP at this
point, the package generates wrapper predicates with internal names
which are compatible with previously hand-written code (e.g.,
shfr_project/5 for the project/5 operation of the 'shfr' domain).

Src-commit: 242f97950194120c12ccc6917f8d4f76fbb2c2df