(ciaopp) fixed shfr handler of builtin ==/2
The previous handler of the builtin X==Y was wrong. It had two clauses
to treat separately the case where either X or Y were (syntactic)
variables, and a third to treat the general case. However those first
two clauses were incorrect, and were also perfectly covered by the
third, so they have been removed. Some other bugs in the third clause
have been fixed too. Some examples of bugs fixed by this are the
following:

a) It was previously inferred bottom for X==Y if X was a variable and
   Y neither ground nor variable (in the code, not at runtime) (e.g.,
   X==f(Y)) because that case was missing in the handler.

b) For the case where X and Y where both variables (in the program,
   not at runtime) and not 'g' in the freeness component, a new
   sharing set was computed (deleting sets with X or Y but not both),
   but the freeness component was not updated accordingly. E.g.,

     :- module(_,[p/0],[]).
     p :- q(X,Y,Z).
     q(X,Y,Z) :- r(X,Y,Z), X==Y.
     r(X,X,a).
     r(X,a,X).

   infers q(X,Y,Z) => ([[X,Y]],[X/f, Y/f, Z/nf]), but should be
   Z/g. If there was Z/f in the calls substitution the handler did
   detect failure however.

c) There was also a bug in the third clause which changed the freeness
   value of some variables to ground although that was not inferred by
   the handler, because the clause which should leave it the same
   failed due to a call to ord_intersect_lists/2 with incorrect
   arguments (e.g., p(A) :- 1==1 would result in ([[A]],[A/g]))

d) There was also a bug which made the analysis fail due to a missing
   catch-all clause with Succ='$bottom' (e.g., p :- X=Y, f(X) == f(a))

Src-commit: cd45b203d32b2119b5811fe435873e06f09aedb5