(ciaopp) added auxinfo to needs/2 in domains
This fixes using ciaopp-dump cmp with det and nf.

needs(AbsInt, aux_info) implies that the information in the abstract
substitutions is not complete and an external solver may be needed.

In practice it is used only to detect whether the typeslib info need to be
included when writing the analysis in a file using ciaopp(p_unit/p_dump).

Src-commit: 7ab40f8df8838031f070601815f442923b92584b