(ciaopp) size analysis in resources omit failing clauses
This commit makes `resources` omit failing clauses, instead of
inferring `bottom` for sizes of output arguments.

Behavior of `fail_clause/1` has changed too: it first searches for
`nf` information about the clause to inspect. If `nf`'s analysis has
not been run, it takes the former approach of searching for a `fail/0`
literal in the clause.

Src-commit: 7e46b0343021a30a6a3173a62ce80092b2b8417f