(ciaopp) Fix test_list_to_mutual_exclusion_test/3 bug
Finish test_list_to_mutual_exclusion_test/3 execution when one of the
clauses has minset test equal to true.

When analyzing:

    :- entry factorial(X,Y) : ( nnegint(X), var(Y) ).

    factorial(X,Y) :- eq(X,0,Comp),
        factorial_(Comp,X,Y).

    factorial_(Comp,_,1) :- Comp = 1.
    factorial_(Comp,X,Y) :- Comp \= 1,
        X1 is X - 1,
        factorial(X1,F),
        Y is X * F.

    eq(X,Y,1) :- X = Y.
    eq(X,Y,0) :- X \= Y.

We find that the tests for `factorial_/3/1` are `true`, but those for
`factorial_/3/2` aren't, so det analysis fails. By finishing the
execution when finding a `true` set of tests, the analysis will not
fail.

Src-commit: e3cb88e51f635cb13db17b349263749fbf32c7af