(ciaopp) Propagate tests from single clause predicates that only call tests in nf
A new flag `Unfold_Tests` is added to nf's abstract substitution. This flag is initialized to `unfold` and set to `not_unfold` when the predicate being analyzed calls a test (directly or not) or it has more than one clause. A predicate will propagate its tests iff this flag is set to `unfold`.

Only tests for not failing clauses are taken into account to perform covering tests.

Src-commit: 5db41d830b961c69d8e10d0c61a4f39b6076825a