(ciaopp) Adapt ctchecks predicates to receive a list of arbitrary domains.
Some predicates received a list of two domains as an argument (expecting a types and a modes domain). Now, a list of arbitrary domains can be handled, solving a bug that prevented `ctchecks_pp` from displaying warnings when using more than two domains. Usage of `none` (meaning no types or sharing domains were being used) are removed as a result.

Src-commit: c116ccbd6bf7b71dc25dfa7b32d85e967eab5ae0