(ciaopp) Always take into account precondition of success assertions
When the precondition of a success assertion cannot be checked, we
cannot assure the assertion fails even when the postcondition
fails. If we have that foo/1 cannot be checked by any domain:

  :- export(p/1).
  :- entry p(X) : int(X).
  p(X) :-
      q(X).

  :- success q(X) : ( int(X), foo(X) ) => var(X).
  q(_).

Incorrect output was:

  :- false success q(X) : ( int(X), foo(X) ) => var(X).

Output status for q/1 used to be false, instead of check. The output
simplified assertion should be:

  :- check success q(X) : foo(X) => fail.

Details:

 - changing the reduce_success rule handling the failing postcondition
   case, we can make this example correct again.

Src-commit: 14b5ff6c0e0171c7dd09e5cb84de51a0db4242c6