(ciaopp) fixed bug in polyhedra:project_on_dimensions/5, adding missing copies
Polyhedra failed when analyzing programs that were not fully numerical (e.g., `p(x) :- X = a.`), due to a wrong implementatin of project (dimensions were removed on the foreign polyhedra side but not on the Prolog vars part of the abstract substitution).

This commit also introduces some missing (foreign) polyhedra copies.

Src-commit: 12b624a5baced33ba3753de49aa11f164249891e