(ciaopp) more depthk fixes
`depth-k` was losing precision when analyzing non-normalized
programs. This was caused by a bug introduced when fixing another bug
in `call_to_success_fact/9` that generated non-idempotent abstractions
(i.e. abstractions where not all the program variables are _only_ in
the left side of the equations).

Instead of using `variables_are_variables/2` to gather the left
variables of the substitution, the union of `Sv` and `Hv` was used
(this is an error because the call to `depthk_unify/3` performed
before can introduce new temporal auxiliary variables that are not in
Sv or Hv). Replacing this by a call to `variables_are_variables/2`
(which gets all the left-side-variables) fix the bug.

Src-commit: 1b437a463652ebe42b72e348821cbd51e3e554fd