(ciaopp,domains) polyhedra infinite loop bug
I added a case in `success_builtin_/5` to check if a constraint contains variables that are not integers. Notice that polyhedra abstraction is a tuple `(Polyhedron, Variables)` where `Variables` are all the variables that **_may be_** integers. Thus, all the variables are initialized as possible integers. Because of that, if a variable does not appear in the list it means that it is not an integer and any constraint (\>, \>=, \<, =\<) will raise an error.

Src-commit: a7185fd055ac2061c628a2cad640de6101b8908b